admin/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Tobias Napoli <tobiasnapoli03@gmail.com>"]
readme = "README.md"
packages = [{include = "web", from = "src"}]

[tool.poetry.dependencies]
python = "3.12.3"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"
flask-session = "^0.8.0"
flask-bcrypt = "^1.0.1"
minio = "^7.2.9"
ulid = "^1.1.0"
flask-wtf = "^1.2.2"
plotly = "^5.24.1"
pandas = "^2.2.3"
marshmallow = "^3.23.1"
flask-cors = "^5.0.0"
authlib = "^1.3.2"
requests = "^2.32.3"
oauthlib = "^3.2.2"
google-auth = "^2.36.0"
google-auth-oauthlib = "^1.2.1"
cachecontrol = "^0.14.1"
email-validator = "^2.2.0"
reportlab = "^4.2.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
